backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
supabase
xgboost
scikit-learn
numpy
pandas
python-dotenv
httpx
slowapi
